+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "like-git-project",
+  "name": "credit-card-form",
   "version": "0.1.0",
   "private": true,
   "scripts": {
