+++ src/App.vue
@@ -1,22 +1,22 @@
 <template>
   <div>
     <b-nav>
-      <b-nav-item>クレジットカード入力フォーム</b-nav-item>
+      <b-nav-item>クレジットカード入力フォームのデモ</b-nav-item>
     </b-nav>
     <b-container class="root-app">
-      <Project />
+      <FormContainer />
     </b-container>
   </div>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
-import Project from "./containers/Project.vue";
+import FormContainer from "./containers/FormContainer.vue";
 
 export default Vue.extend({
   name: "App",
   components: {
-    Project
+    FormContainer
   }
 });
 </script>
